+++ src/Kalnoy/Nestedset/NestedSet.php
@@ -1,75 +1,48 @@
-<?php namespace Kalnoy\Nestedset;
+<?php
 
-use \Illuminate\Database\Connection;
-use \Illuminate\Database\Schema\Blueprint;
+namespace Kalnoy\Nestedset;
+
+use Illuminate\Database\Connection;
+use Illuminate\Database\Schema\Blueprint;
 
 class NestedSet {
 
     /**
-     * Add NestedSet columns to the table. Also create index and foreign key.
+     * Add default nested set columns to the table. Also create an index.
      *
-     * @param   Blueprint  $table
-     *
-     * @return  void
+     * @param \Illuminate\Database\Schema\Blueprint $table
+     * @param string $primaryKey
      */
-    static public function columns(Blueprint $table, $primaryKey = 'id')
+    public static function columns(Blueprint $table, $primaryKey = 'id')
     {
         $table->integer(Node::LFT);
         $table->integer(Node::RGT);
         $table->unsignedInteger(Node::PARENT_ID)->nullable();
 
-        $table->index(array(Node::LFT, Node::RGT, Node::PARENT_ID), 'nested_set_index');
-
-        $table
-            ->foreign(Node::PARENT_ID, self::getForeignKeyName($table->getTable()))
-            ->references($primaryKey)
-            ->on($table->getTable())
-            ->onDelete('cascade');
-    }
-
-    /**
-     * Get foreign key name for the table.
-     *
-     * @param   string  $table
-     *
-     * @return  string
-     */
-    protected static function getForeignKeyName($table)
-    {
-        return $table.'_nested_set_foreign';
+        $table->index(self::getDefaultColumns());
     }
 
     /**
      * Drop NestedSet columns.
      *
-     * @param   Blueprint  $table
-     *
-     * @return  void
+     * @param \Illuminate\Database\Schema\Blueprint $table
      */
-    static public function dropColumns(Blueprint $table)
+    public static function dropColumns(Blueprint $table)
     {
-        $table->dropForeign('nested_set_foreign');
-        $table->dropIndex('nested_set_index');
-        $table->dropColumn(Node::LFT, Node::RGT, Node::PARENT_ID);
+        $columns = self::getDefaultColumns();
+
+        $table->dropIndex($columns);
+        $table->dropColumn($columns);
     }
 
     /**
-     * Create root node.
-     *
-     * @param   string  $table
-     * @param   array   $extra
-     * @param   string  $connection
-     *
-     * @return  boolean
+     * Get a list of default columns.
+     * 
+     * @return array
      */
-    static function createRoot($table, array $extra = array(), $connection = null)
+    public static function getDefaultColumns()
     {
-        $extra = array_merge($extra, array(
-            Node::LFT => 1,
-            Node::RGT => 2,
-            Node::PARENT_ID => NULL,    
-        ));
-
-        return \DB::connection($connection)->table($table)->insert($extra);
+        return [ Node::LFT, Node::RGT, Node::PARENT_ID ];
     }
+
 }